Rename sources and remove PoS namespace [skip ci]
